refactor(#1627): consolidate find_case_invite_for_actor into polling.py
Move _is_case_invite_for and _find_case_invite_for_actor (renamed to
find_case_invite_for_actor) from fcv_demo.py and fvcv_extension_demo.py
into the shared vultron/demo/helpers/polling.py module.

Both scenario files now import the shared helper; the local duplicates
are removed. test_fcv_demo.py updated to import from polling directly.
Closes #1627.
